NativeHub: Reply early in the DSP chain if the sink buffers are full
Once upon a time, a developer had that long ongoing bug that
came out of nowhere. The developer spent months trying to track
it down, blaming other people around him and libraries used.
The developer tried many libraries and many ways of trying to
do it, but eventually, the issue was way trickier than that.

In real terms, DSP effects were glitching every second or so
for unknown reason. After tying multiple Biquad and DSP
effects library, all causing the same behavior, I triple-checked
the full pipeline without any luck. I finally came accross
the idea of dumping the audio that was returned by the
DSP to the main app, and I found that every 3x8192 samples,
you'd have the same 256 or so samples repeated during 8192
samples.
It turns out the providers were doing their jobs, sending
the audio to the main app, then the main app would in turn
forward that to the DSP chain. However, when the DSP would
be done processing the frame, only then the main app would
reply "sink buffers full!" to the provider so it can try again.
This means the DSP effects would process all the retries
from providers, regardless of the sink status, causing
glitches as the biquads were processing the same samples
multiple times instead of the normal audio flow. But besides
the small glitch, it was otherwise inaudible as the overflowing
DSP samples would get rejected by the audio sink just like
regular provider samples, meaning it's inaudible without
processing (if you do a bypass DSP), but glitches can
be heard with biquads as they depend on their previous
values.