Clear cache on unbind (or How to fix a big issue in one line)
Some background: When syncing tracks while the screen is off, the
songs status would never be updated until the app/provider is fully
restarted. While this seemed at first an issue caused by the provider,
it is actually an issue caused by the app. As a matter of fact, if
the screen goes off during sync and nothing is playing, providers
callbacks are removed, but the cache is kept. The tracks sync status
will update in the background, but the cache won't be updated since
the callbacks are removed ; and since the track is in cache, the app
won't re-request it.

By clearing up the cache when a provider disconnects, we can ensure
we're using updated data.