Fix playback service not releasing
Due to android:persistent="true" flag added a few months back,
the payback service wouldn't release anymore. This caused the
plugins to stay in memory until they got killed by Android.

This also fixes a wrong understanding of onBind/onRebind/onUnbind
that caused a wrong counting of the number of connections bound
to the playback service. It is now run using startService and
stopService, which does the intended effect.